+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.0] - 2023-06-26
+
 ### Added
 
 - Add ServiceId annotation [#244](https://github.com/nbbrd/java-service-util/issues/244)
@@ -139,7 +141,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/java-service-util/compare/v1.6.1...HEAD
+[Unreleased]: https://github.com/nbbrd/java-service-util/compare/v1.7.0...HEAD
+[1.7.0]: https://github.com/nbbrd/java-service-util/compare/v1.6.1...v1.7.0
 [1.6.1]: https://github.com/nbbrd/java-service-util/compare/v1.6.0...v1.6.1
 [1.6.0]: https://github.com/nbbrd/java-service-util/compare/v1.5.3...v1.6.0
 [1.5.3]: https://github.com/nbbrd/java-service-util/compare/v1.5.2...v1.5.3